Move second parsing of imagenet bb
Previously, the parsing of the imagnet bb down to only those bounding
boxes corresponding to succesfully downloaded images was built into the
image generator. This meant that it was getting run each time we were
working through the generator, when in reality we only need to run it
once (after we have the final set of images). This parsing (e.g. the
second parsing of the imagenet bounding boxes) is now in its own script.