Update parse_alov_bb.py for easy use in generator
Before this commit, parse_alov_bb.py was built to simply parse all of the bounding box files for the alov videos and store the result in a csv. It is much more helpful and efficient in the image generator that will be used in the net to have the parsing actually set up pairs of the video frames that will be fed into the net. This commit sets that up.